+++ common/src/main/java/net/frozenblock/lib/config/v2/entry/predicates/ConfigPredicate.java
@@ -162,7 +162,7 @@ public interface ConfigPredicate extends Supplier<Boolean> {
 	}
 
 	default PlacementFilter asPlacementFilter() {
-		return new ConfigPlacementFilter<>(this);
+		return new ConfigPlacementFilter(this);
 	}
 
 	default LootItemCondition asLootCondition() {