+++ CHANGELOG.md
@@ -66,3 +66,5 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
   - `fallback`: The fallback value to use if the required Config Entry is not present.
 - Removed `BlockStateRenameFix` and `FrozenEntityRenameFix`, as Vanilla provides the same functionality.
 - Added the `frozenlib:strict_rule_based_state_provider` Block State Provider in 26.3+, an alternate implementation of `minecraft:rule_based_state_provider` that properly accounts for failed nested Rule-Based State Providers.
+- Migrated `StructureGenerationConditionApi` to an event in 26.3+.
+- Migrated `StructurePlacementExclusionApi` to an event in 26.3+.